requirements.txt
# Requirements for AI Automation Application with Nova Act
nova-act
pyinstaller>=6.0.0
python-dotenv
pydantic>=2.0.0

# MCP (Model Context Protocol)
mcp[cli]

# Amazon Bedrock AgentCore (cho browser tool)
bedrock-agentcore

# AWS SDK
boto3>=1.34.0

# Rich console output
rich>=13.0.0

# Fire for CLI
fire>=0.5.0
